Fixed two bugs
(1) Fixed the issue where select box has a memory of the previous selection - i.e., previous selection was not cleared when re-initialize select box. It is resolved by calling clearStore() function for enforce the clear of all items and choices in a select box.
(2) Fixed the issue that download structure function was not working. This is due to an incorrect key used in search.